+++ rmw_gurumdds_cpp/src/rmw_event.cpp
@@ -18,7 +18,7 @@
 
 #include "rmw_gurumdds_cpp/event_converter.hpp"
 #include "rmw_gurumdds_cpp/identifier.hpp"
-#include "rmw_gurumdds_cpp/types.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
 
 static rmw_ret_t
 init_rmw_event(
