+++ rmw_gurumdds_cpp/src/wait.cpp
@@ -1,3 +1,17 @@
+// Copyright 2024 GurumNetworks, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #include <chrono>
 #include <unordered_map>
 #include <unordered_set>
@@ -10,9 +24,9 @@
 
 #include "rmw_gurumdds_cpp/dds_include.hpp"
 #include "rmw_gurumdds_cpp/event_converter.hpp"
-#include "rmw_gurumdds_cpp/types.hpp"
+#include "rmw_gurumdds_cpp/event_info_common.hpp"
+#include "rmw_gurumdds_cpp/event_info_service.hpp"
 #include "rmw_gurumdds_cpp/wait.hpp"
-
 
 #define CHECK_ATTACH(ret) \
   if (ret == dds_RETCODE_OK) { \
